align statistic models to one file
