--- src/app/domains/statistics/altitude-group.model.ts
@@ -1,1 +1,0 @@
-export type AltitudeGroup = 'alt1' | 'alt2' | 'alt3' | 'alt4' | 'alt5';
