--- src/app/domains/statistics/analytics-type.model.ts
@@ -1,1 +1,0 @@
-export type AnalyticsType = 'species' | 'altitude';
