--- src/app/domains/statistics/analytics-value.model.ts
@@ -1,8 +1,0 @@
-export class AnalyticsValue {
-  phenophase: string;
-  max: Date;
-  median: Date;
-  min: Date;
-  quantile_25: Date;
-  quantile_75: Date;
-}
