+++ src/app/shared/services/form-persistence.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import { AnalyticsType } from '../../domains/statistics/analytics-type.model'; //fixme
+import { AnalyticsType } from '../../domains/statistics/statistics.model'; //fixme
 import { SourceFilterType } from '@shared/models/source-type.model';
 
 @Injectable({ providedIn: 'root' })
